[Issue #1527] Move the "Follow you" badge into the profile header and out-of-line with the often long and already space-constrained username/display name text
